Bump Expression.numSubexpressions from UInt8 to UInt
We've seen crashes due to it overflowing the UInt8 range and the
Expression struct likely anyway has unused padding bytes that can be
used for the integer instead.

+++ Sources/Fuzzilli/Lifting/Expression.swift
@@ -52,7 +52,7 @@ public struct Expression: CustomStringConvertible {
     public let type: ExpressionType
     public let text: String
 
-    let numSubexpressions: UInt8
+    let numSubexpressions: UInt
 
     public var characteristic: ExpressionType.Characteristic {
         return type.characteristic
